Refactor (almost) all events into "Source" "Action" "Target".
This properly restructures and renames all remaining events to
follow "SourceActionTargetEvent" naming scheme.

Additionally, this makes significant changes to snapshotting of
BlockStates and Entity instances where the snapshots are
ImmutableDataHolders. Likewise, for ChangeBlockEvent, an additional
"helper" class is added: BlockTransaction. A BlockTransaction
keeps track of the original BlockSnapshot that is being swapped,
the default BlockSnapshot replacement (basically what vanilla says),
and the custom BlockSnapshot provided by plugins through events.
Conversely, the BlockTransaction can be marked as invalid to
prevent the change of the specific transaction, further making
ChangeBlockEvent#getTransactions() an ImmutableList of BlockTransactions.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>